split up database/mod.rs
