+++ src/app.rs
@@ -11,7 +11,8 @@ use tracing::{info, instrument, warn};
 
 use crate::config::Config;
 use crate::contracts::{IdentityManager, SharedIdentityManager};
-use crate::database::{Database, DatabaseExt as _};
+use crate::database::query::DatabaseQuery as _;
+use crate::database::Database;
 use crate::ethereum::Ethereum;
 use crate::identity_tree::{
     CanonicalTreeBuilder, Hash, InclusionProof, ProcessedStatus, RootItem, Status, TreeState,
